change approach fetch for microcms
