+++ package.json
@@ -14,6 +14,7 @@
     "babel-core": "^6.26.3",
     "eslint-config-next": "latest",
     "gsap": "^3.6.0",
+    "microcms-js-sdk": "^3.1.1",
     "next": "latest",
     "next-images": "^1.7.0",
     "react": "latest",
