+++ src/components/BaseHeader.tsx
@@ -1,29 +1,16 @@
-import React from "react";
-
-interface titleType {
+interface Props {
   title: string;
 }
 
-class BaseHeader extends React.Component<titleType> {
-  constructor(props: titleType) {
-    super(props);
-    this.state = {
-      title: "BaseHeader",
-    };
-  }
-
-  render() {
-    return (
-      <header className="base-header">
-        <div className="header--body">
-          <div className="logo">
-            {/* <img src={logo} className="App-logo" alt="logo" /> */}
-          </div>
-          <h3 className="header__title">{/* { this.props.title } */}</h3>
+export default function BaseHeader(props: Props) {
+  return (
+    <header className="base-header">
+      <div className="header--body">
+        <div className="logo">
+          {/* <img src={logo} className="App-logo" alt="logo" /> */}
         </div>
-      </header>
-    );
-  }
+        <h3 className="header__title">{/* { props.title } */}</h3>
+      </div>
+    </header>
+  )
 }
-
-export default BaseHeader;
