+++ src/libs/microcms.client.ts
@@ -1,0 +1,6 @@
+import { createClient } from 'microcms-js-sdk';
+
+export const client = createClient({
+  serviceDomain: "nextfield", 
+  apiKey: process.env.NEXT_PUBLIC_MICROAPI_KEY,
+});
